requirements.txt
torch-neuronx>=2.1
neuronx-cc
transformers>=4.40.0,<4.45.0
einops
timm
pillow
requests

# vLLM server dependencies
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing dependencies
pytest>=7.0.0
hypothesis>=6.0.0
httpx>=0.24.0  # For TestClient
